requirements.txt
flask
flask_sqlalchemy
sqlalchemy
gunicorn
psycopg2
mysql-connector
mysql-connector-python
pymysql
cryptography